The "adduser" management command is now "users".
The "users" management command will allow you to add and edit users.
Documentation and UPDATING reflect this change.
