+++ README.md
@@ -164,11 +164,11 @@ Run the `create_default_collections` management command to setup your database:
 
 ### Add your first user:
 
-Take a look at the options for the **adduser** management command:
+Take a look at the options for the **users** management command:
 
 ```bash
 
-    python manage.py adduser -h
+    python manage.py users -h
 ```
 
 Use that command to setup your first admin user for CRITs. Be sure to use **-a** to set them as an admin. **Make note of the temporary password provided in the output!!**
@@ -351,11 +351,11 @@ Run the `create_default_collections` management command to setup your database:
 
 ### Add your first user:
 
-Take a look at the options for the adduser management command:
+Take a look at the options for the **users** management command:
 
 ```bash
 
-    python manage.py adduser -h
+    python manage.py users -h
 ```
 
 Use that command to setup your first admin user for CRITs. Be sure to use **-a** to set them as an admin. **Make note of the temporary password provided in the output!!**
